Relocating Input Feeder class
