+++ modules/siddhi-launcher/src/main/java/org/wso2/siddhi/launcher/debug/VMDebugManager.java
@@ -19,21 +19,18 @@
 package org.wso2.siddhi.launcher.debug;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.gson.Gson;
 import io.netty.channel.Channel;
-import org.wso2.siddhi.core.SiddhiAppRuntime;
 import org.wso2.siddhi.core.SiddhiManager;
 import org.wso2.siddhi.launcher.debug.dto.CommandDTO;
 import org.wso2.siddhi.launcher.debug.dto.MessageDTO;
 import org.wso2.siddhi.launcher.exception.DebugException;
 import org.wso2.siddhi.launcher.exception.SiddhiException;
 import org.wso2.siddhi.launcher.internal.DebugRuntime;
+import org.wso2.siddhi.launcher.util.InputFeeder;
 import org.wso2.siddhi.launcher.util.PrintInfo;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.Scanner;
-import java.util.concurrent.atomic.AtomicBoolean;
 
 /**
  * {@code VMDebugManager} Manages debug sessions and handle debug related actions.
@@ -259,105 +256,5 @@ public class VMDebugManager {
         message.setCode(DebugConstants.CODE_ACK);
         message.setMessage(messageText);
         debugServer.pushMessageToClient(debugSession, message);
-    }
-
-    /**
-     * Method name used to specify delay in input.
-     * Example: delay(100)
-     */
-    private static final String DELAY = "delay";
-
-    /**
-     * Delimiter separating stream name and the input data in the input.
-     */
-    private static final String INPUT_DELIMITER = "=";
-
-    /**
-     * A runnable class to feed the input to the Siddhi runtime.
-     */
-    private static class InputFeeder implements Runnable {
-        private final SiddhiAppRuntime siddhiAppRuntime;
-        private String input;
-        private volatile AtomicBoolean running = new AtomicBoolean(false);
-        private Thread thread;
-
-        private InputFeeder(SiddhiAppRuntime siddhiAppRuntime, String input) {
-            this.siddhiAppRuntime = siddhiAppRuntime;
-            this.input = input;
-        }
-
-        @Override
-        public void run() {
-            // Scanner to read the user input line by line
-            Scanner scanner = new Scanner(input);
-            Gson gson = new Gson();
-            while (scanner.hasNext()) {
-                if (!running.get()) {
-                    break;
-                }
-                String line = scanner.nextLine().trim();
-                if (line.startsWith(DELAY)) {
-                    // The delay(<time in milliseconds>) is used to delay the input
-                    line = line.substring(6, line.length() - 1);
-                    try {
-                        Thread.sleep(Integer.parseInt(line));
-                    } catch (InterruptedException e) {
-                        PrintInfo.error("Error in waiting for " + line + " milliseconds");
-                    }
-                } else {
-                    // The inout format is: <stream name>=<data in json object[] format>
-                    String[] components = line.split(INPUT_DELIMITER);
-                    String streamName = components[0];
-                    String event = components[1];
-                    Object[] data = gson.fromJson(event, Object[].class);
-                    PrintInfo.info("@Send: Stream: " + streamName + ", Event: " + event);
-                    try {
-                        siddhiAppRuntime.getInputHandler(streamName).send(data);
-                    } catch (InterruptedException e) {
-                        PrintInfo.error("Error in sending event " + event + " to Siddhi");
-                    }
-                }
-            }
-            scanner.close();
-        }
-
-        /**
-         * Check whether the input feeder is running or not.
-         *
-         * @return true if the input feeder is running or not stopped manually, otherwise false.
-         */
-        boolean isRunning() {
-            return running.get();
-
-        }
-
-        /**
-         * Stop the input feeder.
-         */
-        void stop() {
-            this.running.set(false);
-        }
-
-        /**
-         * Start the input feeder.
-         */
-        public void start() {
-            if (!this.running.get()) {
-                this.running.set(true);
-                thread = new Thread(this);
-                thread.start();
-            }
-        }
-
-        /**
-         * Join the current thread behind the thread used to execute the input feeder.
-         */
-        public void join() {
-            try {
-                this.thread.join();
-            } catch (InterruptedException e) {
-                PrintInfo.error("Error in joining the main thread behind the input feeder");
-            }
-        }
     }
 }
